chore: Move feature state to jotai
