+++ contexts/index.ts
@@ -1,3 +1,2 @@
 export { default as ProgressProvider } from './ProgressContext'
 export { default as DataProvider } from './DataContext'
-export { default as FeatureProvider } from './FeatureProvider'
