+++ types/interfaces.ts
@@ -83,12 +83,6 @@ export interface DataContextType {
   setCurrentLanguage: (language: Language) => void
 }
 
-export interface FeatureContextType {
-  isLoading: boolean
-  features: { [key: string]: number }
-  isFeatureEnabled: (key: string) => boolean
-}
-
 export interface ModalState {
   open: boolean
   meta: any
